requirements.txt
nltk==3.8.1
pandas==2.1.3
Pillow==9.4.0
Pillow==10.1.0
scikit_learn==1.3.2
streamlit==1.22.0